Reorganize deploy and load into conda dir
Update create_compass_env.py to work with py2.7 (before
bootstrapping).

Reorganize script into more functions.

Update script for new directory structure.
