+++ conda/compass_env/build.template
