+++ conda/compass_env/load_compass.template
