+++ conda/compass_env/spec-file-mpich.txt
