+++ conda/compass_env/spec-file-nompi.txt
@@ -31,12 +31,6 @@ scipy
 xarray
 
 # Development
-c-compiler
-cmake
-cxx-compiler
-fortran-compiler
-m4
-netcdf-fortran
 pip
 
 # Documentation
