+++ conda/compass_env/spec-file-openmpi.txt
