--- recipe/LICENSE
@@ -1,39 +1,0 @@
-Copyright (c) 2013-2020,  Los Alamos National Security, LLC (LANS) (Ocean: LA-CC-13-047;
-Land Ice: LA-CC-13-117) and the University Corporation for Atmospheric Research (UCAR).
-
-All rights reserved.
-
-LANS is the operator of the Los Alamos National Laboratory under Contract No.
-DE-AC52-06NA25396 with the U.S. Department of Energy.  UCAR manages the National
-Center for Atmospheric Research under Cooperative Agreement ATM-0753581 with the
-National Science Foundation.  The U.S. Government has rights to use, reproduce,
-and distribute this software.  NO WARRANTY, EXPRESS OR IMPLIED IS OFFERED BY
-LANS, UCAR OR THE GOVERNMENT AND NONE OF THEM ASSUME ANY LIABILITY FOR THE USE
-OF THIS SOFTWARE.  If software is modified to produce derivative works, such
-modified software should be clearly marked, so as not to confuse it with the
-version available from LANS and UCAR.
-
-Additionally, redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-
-1) Redistributions of source code must retain the above copyright notice, this
-list of conditions and the following disclaimer.
-
-2) Redistributions in binary form must reproduce the above copyright notice,
-this list of conditions and the following disclaimer in the documentation and/or
-other materials provided with the distribution.
-
-3) None of the names of LANS, UCAR or the names of its contributors, if any, may
-be used to endorse or promote products derived from this software without
-specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
-ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
-ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
