--- recipe/build_and_upload.bash
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-conda build .
-
-#anaconda upload -u e3sm ${HOME}/miniconda3/conda-bld/noarch/compass*.tar.bz2
